# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ecosystem / EnvTrack API
  version: 1.0.0
  description: |
    Environmental reporting backend (Go + Gin + PostgreSQL).

    ## Authentication
    Most endpoints require a Bearer JWT obtained from `POST /api/login`.
    Include the header: `Authorization: Bearer <access_token>`.

    ## Response envelope
    - **Success:** `{ "success": true, "data": <payload>, "message": "<string>" }`
    - **Error:** `{ "success": false, "error": "<string>", "code": <http_status> }`

    ## Interactive docs
    Open [Swagger UI](/api/docs) in this server.
servers:
  - url: http://localhost:3030
    description: Local / Docker backend
...
paths:
...
  /api/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      operationId: forgotPassword
      description: Always returns success message (no email enumeration). Sends 1h JWT reset link if account exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset email sent if account exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSuccess"
  /api/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with JWT from email
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/change-password:
    post:
      tags: [Auth]
      summary: Change password (authenticated)
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/profile:
    get:
      tags: [Profile]
      summary: Get current user profile
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Profile]
      summary: Delete own account (soft delete)
      operationId: deleteAccount
      security:
        - bearerAuth: []
      description: Sets deleted_at and status='deleted'. Logs audit entry. Blacklists current token.
      responses:
        "200":
          description: Account deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/login
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: invalid category
            code: 400
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: Invalid or expired token
            code: 401
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: Report not found
            code: 404
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: Database error
            code: 500
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          description: Endpoint-specific payload
        message:
          type: string
    ErrorEnvelope:
      type: object
      required: [success, error, code]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: integer
          description: HTTP status code duplicated in body
    MessageSuccess:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              nullable: true
              example: null
    UserRole:
      type: string
      enum: [citizen, admin, authority]
    UserStatus:
      type: string
      enum: [active, suspended, banned, deleted]
...
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
          description: JWT from reset email (purpose password_reset, 1h)
        new_password:
          type: string
          format: password
          minLength: 6
    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
          minLength: 6
    User:
      type: object
      properties:
        user_id:
          type: integer
        email:
          type: string
          format: email
        phone_number:
          type: string
          example: "+243812345678"
        first_name:
          type: string
        last_name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        is_email_verified:
          type: boolean
        reputation_score:
          type: integer
          description: "+10 investigating approval, +1 upvote received, -5 duplicate"
        total_reports:
          type: integer
        status:
          $ref: "#/components/schemas/UserStatus"
        created_at:
          type: string
          format: date-time
...